requirements.txt
fastapi[standard]
sqlmodel
pyjwt
pwdlib[argon2]
pydantic_settings